teachmeskills/FastAPI_aiogram/aiogram_bot/bot.py
import os
from aiogram import Bot, Dispatcher
from aiogram.fsm.storage.memory import MemoryStorage
from dotenv import load_dotenv

# Load environment variables from a .env file
load_dotenv()

# Initialize the Bot with the token retrieved from environment variables
bot = Bot(token=os.getenv("TELEGRAM-TOKEN"))

# Initialize the Dispatcher to manage updates and handlers
dp = Dispatcher(storage=MemoryStorage())
